requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
mysql-connector-python==8.0.20
protobuf==3.11.3
requests==2.23.0
six==1.15.0
urllib3==1.25.9
